chore: fix requests content type and rest api responses

+++ includes/importers/class-themeisle-ob-content-importer.php
@@ -20,7 +20,12 @@ class Themeisle_OB_Content_Importer {
 	 */
 	public function import_remote_xml( WP_REST_Request $request ) {
 		if ( ! current_user_can( 'manage_options' ) ) {
-			wp_send_json_error( 'ti__ob_permission_err_1', 500 );
+			return new WP_REST_Response(
+				array(
+					'data'    => 'ti__ob_permission_err_1',
+					'success' => false,
+				)
+			);
 		}
 
 		do_action( 'themeisle_ob_before_xml_import' );
@@ -30,11 +35,21 @@ class Themeisle_OB_Content_Importer {
 		$content_file_url = $body['contentFile'];
 
 		if ( empty( $content_file_url ) ) {
-			wp_send_json_error( 'ti__ob_remote_err_1', 500 );
+			return new WP_REST_Response(
+				array(
+					'data'    => 'ti__ob_remote_err_1',
+					'success' => false,
+				)
+			);
 		}
 
 		if ( ! isset( $body['source'] ) || empty( $body['source'] ) ) {
-			wp_send_json_error( 'ti__ob_remote_err_2', 500 );
+			return new WP_REST_Response(
+				array(
+					'data'    => 'ti__ob_remote_err_2',
+					'success' => false,
+				)
+			);
 		}
 
 		set_time_limit( 10000 );
@@ -52,7 +67,16 @@ class Themeisle_OB_Content_Importer {
 			$content_file_path = $content_file_url;
 		}
 
-		$this->import_file( $content_file_path );
+		$import_status = $this->import_file( $content_file_path );
+
+		if ( is_wp_error( $import_status ) ) {
+			return new WP_REST_Response(
+				array(
+					'data'    => $import_status,
+					'success' => false,
+				)
+			);
+		}
 
 		if ( $body['source'] === 'remote' ) {
 			unlink( $content_file_path );
@@ -76,12 +100,17 @@ class Themeisle_OB_Content_Importer {
 		do_action( 'themeisle_ob_after_shop_pages_setup' );
 
 		if ( empty( $frontpage_id ) ) {
-			wp_send_json_error( 'ti__ob_front_page_id_err_1', 500 );
+			return new WP_REST_Response(
+				array(
+					'data'    => 'ti__ob_front_page_id_err_1',
+					'success' => false,
+				)
+			);
 		}
 
-		wp_send_json_success(
+		return new WP_REST_Response(
 			array(
-				'message'      => 'Success',
+				'success'      => true,
 				'frontpage_id' => $frontpage_id,
 			)
 		);
@@ -195,7 +224,7 @@ class Themeisle_OB_Content_Importer {
 	 */
 	private function import_file( $file_path ) {
 		if ( empty( $file_path ) || ! file_exists( $file_path ) || ! is_readable( $file_path ) ) {
-			wp_send_json_error( 'ti__ob_content_err_1', 500 );
+			return new WP_Error( 'ti__ob_content_err_1' );
 		}
 		$this->load_importer();
 
@@ -204,9 +233,8 @@ class Themeisle_OB_Content_Importer {
 
 		$importer = new Themeisle_OB_WXR_Importer();
 		$result   = $importer->import( $file_path );
-		if ( is_wp_error( $result ) ) {
-			wp_send_json_error( $result, 500 );
-		}
+
+		return $result;
 	}
 
 	/**
